+++ src/constants/LocalStorageKey.js
@@ -1,0 +1,1 @@
+export const LOCAL_STORAGE_KEY = 'lunch_restaurants';
